requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
pandas==2.2.3
numpy==2.2.6
scikit-learn==1.6.1
pydantic==2.11.5
python-multipart==0.0.20
httpx==0.28.1
pytest==8.3.5
pytest-asyncio==1.0.0
black==25.1.0
flake8==7.2.0
mypy==1.15.0
isort==6.0.1
sqlalchemy==2.0.41
redis==6.2.0
aioredis==2.0.1
slowapi==0.1.9
psycopg[binary]==3.2.9
python-dotenv==1.0.0
torch==2.7.0
sentence-transformers==4.1.0